backend/requirements.txt
python-dotenv
imbox
openai
langchain
